ACN_backend_service/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.2
gunicorn==20.1.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.2.0
psycopg2==2.8.6
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
wrapt==1.12.1
